[♻️ refactor] Mainpage code refactoring
