+++ src/App.js
@@ -5,7 +5,7 @@ import Homepage from './components/Homepage';
 function App() {
   return (
     <div className="App">
-      <Homepage></Homepage>
+      <Homepage></Homepage>  {/* 첫 메인페이지 */}
     </div>
   );
 }
