+++ src/components/Homepage.js
@@ -1,19 +1,19 @@
 import * as React from 'react';
-import AppFooter from './modules/views/AppFooter';
-import MainpageBox from './modules/views/MainpageBox';
-import AppBar from './modules/views/AppBar';
-import withRoot from './modules/withRoot';
+import AppFooter from './views/AppFooter';
+import MainpageBox from './views/MainpageBox';
+import AppBar from './views/AppBar';
+import withRoot from './withRoot';
 
 function Index() {
   return (
     <React.Fragment>
+      {/* 
+        AppBar: 맨 위 상단바 구현
+        MainpageBox: 가운데 이미지박스 + 학교 셀렉트 부분 구현
+        AppFooter: 페이지 하단바 구현
+      */}
       <AppBar></AppBar>
       <MainpageBox></MainpageBox>
-      {/* <ProductValues />
-      <ProductCategories />
-      <ProductHowItWorks />
-      <ProductCTA />
-      <ProductSmokingHero /> */}
       <AppFooter />
     </React.Fragment>
   );
