+++ src/components/assets/Paper.js
