+++ src/components/views/AppBar.js
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import Box from '@mui/material/Box';
 import Link from '@mui/material/Link';
-import AppBar from '../components/AppBar';
-import Toolbar from '../components/Toolbar';
+import AppBar from '../assets/AppBar';
+import Toolbar from '../assets/Toolbar';
 
 const rightLink = {
   fontSize: 16,
@@ -12,14 +12,17 @@ const rightLink = {
 const LogoImage =
   'https://user-images.githubusercontent.com/82520143/161506667-c500b236-f034-4d85-b3b2-b16933d7403a.png';
 
+// ./asset 에서 레이아웃 틀을 가져다가 사용함
 function AppAppBar() {
   return (
     <div>
       <AppBar
-        style={{background: '#24527a'}} 
+        style={{background: '#24527a'}} // 배경색 지정
         position="fixed"
       >
+        {/* 팀 로고 이미지 & 텍스트 배치 */}
         <Toolbar sx={{ justifyContent: 'space-between' }}>
+          {/* 로고 이미지 */}
           <Box
             component="img"
             sx={{
@@ -28,7 +31,7 @@ function AppAppBar() {
             alt="Logo"
             src={LogoImage}
           />
-
+          {/* 텍스트 배치 */}
           <Link
             variant="h6"
             underline="none"
