+++ src/components/views/MainpageBoxLayout.js
@@ -5,7 +5,7 @@ import { styled } from '@mui/material/styles';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
 
-const ProductHeroLayoutRoot = styled('section')(({ theme }) => ({
+const MainPageBoxLayoutRoot = styled('section')(({ theme }) => ({
   color: theme.palette.common.white,
   position: 'relative',
   display: 'flex',
@@ -28,11 +28,11 @@ const Background = styled(Box)({
   zIndex: -2,
 });
 
-function ProductHeroLayout(props) {
+function MainPageBoxLayout(props) {
   const { sxBackground, children } = props;
 
   return (
-    <ProductHeroLayoutRoot>
+    <MainPageBoxLayoutRoot>
       <Container
         sx={{
           mt: 3,
@@ -58,11 +58,11 @@ function ProductHeroLayout(props) {
         <Background sx={sxBackground} />
       
       </Container>
-    </ProductHeroLayoutRoot>
+    </MainPageBoxLayoutRoot>
   );
 }
 
-ProductHeroLayout.propTypes = {
+MainPageBoxLayout.propTypes = {
   children: PropTypes.node,
   sxBackground: PropTypes.oneOfType([
     PropTypes.arrayOf(
@@ -73,4 +73,4 @@ ProductHeroLayout.propTypes = {
   ]),
 };
 
-export default ProductHeroLayout;
+export default MainPageBoxLayout;
